This is my first commit. I have added the basic requirements of the demo including movement using WASD, collecting the pickups and triggering an exit flag which will allow one to complete the game.
I will next work on refining the demo and cleaning up and left over code. I have a list of features that I would like to try and get completed such as an Achievements system, Gamepad , scriptable objects and some unit tests.

Currently the pickup items I have added don't refer to any game theme ideas that I have, but I will plan a design and build a level around them.

There are areas I would to improve such as a better system for handling  the events, I'm currently using a singleton pattern which will be fine for the scope of this demo, but I feel that the design could be improved. I'm going down to London during the week so will prioritise this if I have any time left.